requirements.txt
torch>=1.0
torchvision>=0.2.1
tqdm
numpy
pydicom
SimpleITK
Pillow
PyYAML
opencv-python
pygraphviz
graphviz
scipy
scikit-image
adabound
tensorboardX